+++ src/test/java/shop/S5G/shop/ShopApplicationTests.java
@@ -9,7 +9,5 @@ class ShopApplicationTests {
 	@Test
 	void contextLoads() {
 	}
-	@Test
-	void testThenReturn() {
-	}
+
 }
