+++ src/test/java/shop/S5G/shop/controller/order/OrderControllerTest.java
@@ -24,7 +24,6 @@ import org.testcontainers.shaded.com.fasterxml.jackson.databind.ObjectMapper;
 import shop.S5G.shop.dto.order.OrderCreateResponseDto;
 import shop.S5G.shop.dto.order.OrderWithDetailResponseDto;
 import shop.S5G.shop.exception.member.CustomerNotFoundException;
-import shop.S5G.shop.exception.order.OrderDetailsNotExistException;
 import shop.S5G.shop.service.order.OrderDetailService;
 import shop.S5G.shop.service.order.OrderService;
 
@@ -71,29 +70,6 @@ class OrderControllerTest {
         verify(orderService, times(1)).queryAllOrdersByCustomerId(3L);
     }
 
-    @Test
-    void fetchOrderDetailsEmptyTest() throws Exception{
-        when(orderDetailService.findOrderDetailsByOrderId(anyLong())).thenReturn(List.of());
-
-        mvc.perform(MockMvcRequestBuilders.get("/api/shop/orders/1"))
-            .andExpect(status().isOk())
-            .andExpect(content().string(org.hamcrest.Matchers.equalTo("[]")));
-
-        verify(orderDetailService, times(1)).findOrderDetailsByOrderId(1L);
-    }
-
-    @Test
-    void fetchOrderDetailsErrorTest() throws Exception {
-        when(orderDetailService.findOrderDetailsByOrderId(anyLong())).thenThrow(
-            new OrderDetailsNotExistException("OrderDetails do not exist")
-        );
-
-        mvc.perform(MockMvcRequestBuilders.get("/api/shop/orders/1"))
-            .andExpect(status().isNotFound())
-            .andExpect(content().string(containsString("not exist")));
-
-        verify(orderDetailService, times(1)).findOrderDetailsByOrderId(1L);
-    }
     String validatedTestCase = """
         {
             "customerId": 1,
